% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcelmer.R
\docType{package}
\name{psrcelmer}
\alias{psrcelmer-package}
\alias{psrcelmer}
\title{psrcelmer: A package for easy retrieval of data from PSRC's databases Elmer and ElmerGeo}
\description{
Provides streamlined methods to retrieve data from PSRC's databases Elmer and ElmerGeo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psrc.github.io/psrcelmer/}
}

}
\author{
Chris Peak \email{cpeak@psrc.org}

Mike Jensen \email{mjensen@psrc.org}

\bold{Homepage}: \url{https://psrc.github.io/psrcelmer/}
}
